:bug: Fix: Add deployedBytecode for runtime scripting (#690)
## Description

Bytecode and deployedBytecode are different. Bytecode is for deploying
deployedBytecode is for running. EVMts vm expects deployedBytecode so we
need to add it to the EVMts contract in @evmts/core and also make
@evmts/bundler resolve deployedBytecode

## Testing

Explain the quality checks that have been done on the code changes

## Additional Information

- [ ] I read the [contributing docs](../docs/contributing.md) (if this
is your first contribution)

Your ENS/address:

Co-authored-by: Will Cory <willcory@Wills-MacBook-Pro.local>