:sparkles: Feat: Add eth json-rpc to api (#819)
## Description

Add all eth_ JSON-RPC methods to APi. We need this to be able to respond
to JSON-RPC requests

Results are just returning the Response type atm and will need to be
updated (e.g. replace hex with bigint) as handlers are implemented

## Testing

Explain the quality checks that have been done on the code changes

## Additional Information

- [ ] I read the [contributing docs](../docs/contributing.md) (if this
is your first contribution)

Your ENS/address:



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
	- Enhanced support for Ethereum blockchain operations in JSON-RPC API.
- Introduced new response types for Ethereum-related JSON-RPC
procedures.
- Added new types for blockchain event filtering and transaction
parameters.

- **Improvements**
- Updated existing types to enforce immutability of blockchain data
representations.

- **Documentation**
- Added comments to clarify the purpose of specific JSON-RPC response
types.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Will Cory <willcory@Wills-MacBook-Pro.local>