fix: allow multiple topics for eth_getLogs and eth_getFilterLogs (#1473)
https://ethereum.org/en/developers/docs/apis/json-rpc/#eth_getlogs

```
topics: Array of DATA, - (optional) Array of 32 Bytes DATA topics. Topics are order-dependent. Each topic can also be an array of DATA with "or" options.
```

each topic can be an array of Data. The correct API was being used for
receiptManager.getLogs, so just needed to update eth_getLogs and
eth_getFilterLogs handler to correctly pass it

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Enhanced log filtering capabilities by allowing multiple topics in log
queries.
  
- **Documentation**
- Improved clarity and accuracy in documentation, including updates to
descriptions and hyperlink references.

- **Tests**
- Added new test cases to validate the functionality of log filtering
with multiple topics.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Will Cory <willcory10@gmail.com>