refactor: make mud plugin work correctly (#1729)
The MUD plugin before that has several flaws, and is super slow to
execute optimistic updates, which makes the whole plugin not really
relevant.

This PR introduces multiple fixes to highly reduce the occurrence of
sync inconsistencies/race conditions, some layer of caching and
optimizations, and attempts to fix remaining issues.

Fixes:
- Before, when getting the optimistic state, for each tx ran, we would
apply the logs and notify subscribers, which would cause consumers to
request the state, hence run that tx again. This would be problematic as
multiple txs need to be processed, since each tx update would trigger
another run with all the txs so far again. The fix here is simply to run
txs and apply the logs to build up the state, and notify subscribers
only after we ran all the txs.
- Much better separation between view and mutable state. This causes the
main optimistic state getters to become sync, as we now "cache" the
state.

Two biggest pain points that make the whole plugin broken and
unoptimized respectively, are the following:
- [x] There is an architectural sync issue: we have two separate,
asynchronous event streams that should be synchronized but aren't:
- canonical sync stream: stash updates when canonical transactions are
indexed
- optimistic pool stream: txremoved events when optimistic transactions
are removed
These streams are inherently out of sync because:
- The canonical sync happens when the indexer processes blocks
- The optimistic tx removal happens when we detect the canonical tx
exists
- We don't have the canonical tx hash available when the sync event
fires
**-> For now, this one has been fixed by appending a small hex string to
the end of both the canonical and optimistic txs data so we can link
them back together later**
- [ ] Everytime we run a call locally basically, we're clearing the fork
cache for the store address, as if we didn't, it would directly retrieve
storage values from that cache and never hit our getStorageAt override
(that fetches from the synced MUD store instead). This is super bad
design, and also causes a bunch of other storage slots to be
invalidated, when these actually should stay.
- [ ] We'll probably want our fork clients to rebase frequently so it
considers the latest fork state when executing txs